compat: fix clang-tidy warnings
The organization here of the headers was a mess, because the headers
only needed to work out at the end when included in `run.cc`. However,
bazel runs clang-tidy on header only libraries, so we actually need
each header to compile standalone.

(cherry picked from commit 7cc7d0025b7fb7e2836ff348150c6c10f678b27d)
