+++ src/v/compat/begin_group_tx_compat.h
@@ -14,6 +14,7 @@
 #include "compat/begin_group_tx_generator.h"
 #include "compat/check.h"
 #include "compat/json.h"
+#include "compat/model_json.h"
 
 namespace compat {
