+++ src/v/compat/commit_tx_generator.h
@@ -10,7 +10,7 @@
  */
 #pragma once
 
-#include "cluster/types.h"
+#include "cluster/tx_protocol_types.h"
 #include "compat/generator.h"
 #include "compat/tx_gateway_generator.h"
 #include "model/tests/randoms.h"
