+++ src/v/compat/get_node_health_compat.h
@@ -13,6 +13,7 @@
 
 #include "cluster/types.h"
 #include "compat/check.h"
+#include "compat/cluster_json.h"
 #include "compat/get_node_health_generator.h"
 #include "compat/json.h"
